Vector2 moved into separate module
